requirements.txt
absl-py==2.3.0
altair==4.2.2
astunparse==1.6.3
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
choreographer==1.0.9
click==8.1.8
cmdstanpy==1.2.5
colorama==0.4.6
contourpy==1.3.0
convertdate==2.4.0
cycler==0.12.1
defusedxml==0.7.1
entrypoints==0.4
ephem==4.2
et_xmlfile==2.0.0
flatbuffers==25.2.10
fonttools==4.58.4
fpdf2==2.8.3
gast==0.4.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.73.0
h5py==3.14.0
holidays==0.75
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kaleido==0.2.1
keras==2.13.1
kiwisolver==1.4.7
libclang==18.1.1
logistro==1.1.0
LunarCalendar==0.0.9
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mdurl==0.1.2
numpy==1.24.3
oauthlib==3.3.1
openpyxl==3.1.2
opt_einsum==3.4.0
orjson==3.10.18
packaging==25.0
pandas==2.0.3
patsy==1.0.1
pillow==11.2.1
plotly==5.18.0
prophet==1.1.4
protobuf==3.20.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydeck==0.9.1
Pygments==2.19.2
PyMeeus==0.5.12
Pympler==1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
pywin32==310
referencing==0.36.2
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.0.0
rpds-py==0.25.1
rsa==4.9.1
scikit-learn==1.3.0
scipy==1.10.1
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
stanio==0.5.1
statsmodels==0.14.0
streamlit==1.22.0
tenacity==8.5.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl==3.6.0
toml==0.10.2
toolz==1.0.0
tornado==6.5.1
tqdm==4.67.1
typing_extensions==4.5.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
validators==0.35.0
watchdog==6.0.0
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.23.0
